+++ bobot/Response.py
@@ -1,138 +1,45 @@
 """
-    User response module
+    User responses module.
+    Contains:
+        Message
+        Text
+        Keyboard
+        Location
+        File
+        Photo
+        Contact
 """
 
-from bobot.utils.utils import flatten
-from bobot.Errors import RuleNameError, ResponseFormatError, ResponseMessageError
+from bobot.utils.utils import instanceof
+from bobot.Errors import ResponseFormatError
 
-def createMessage(message, data):
-    """
-        Creates message from message description object
-        and update's data
-    """
-    options = {}
-    text = message
+class Message:
+    "Main Message class container"
 
-    if isinstance(message, dict):
-        text = message.get('text')
-        options = message.get('options', {})
+    method = 'sendMessage'
 
-        if not text:
-            raise ResponseMessageError('Message description should have "text" [String] property', 'text')
+    def __init__(self, **options):
+        self.options = options
 
-        if message.get('interpolate', False):
-            text = text.format(
-                text=data.get('text'),
-                first_name=data.get('first_name'),
-                username=data.get('username'),
-                second_name=data.get('second_name'))
+    def getData(self, data):
+        # pylint: disable=unused-argument
+        "Return data to send with method"
 
-    return text, options
+        return []
 
-def processResponseBody(responseBody):
-    "Processing response description body to actions lists"
+    def getOptions(self):
+        "Return options dict"
 
-    if isinstance(responseBody, dict):
-        return [responseBody]
-    elif isinstance(responseBody, list):
-        return list(map(processResponseBody, responseBody))
-    else:
-        raise ResponseFormatError('Wrong response description. Expected dict, got {}'.format(type(responseBody)))
+        options = self.options.get('options', {})
+        options.update({
+            'disable_notifications': self.options.get('sielent', False)
+        })
 
-def sendMessage(message=''):
-    "Returns message sending function"
+        return options
 
-    def action(bot, chatId, data):
-        "Sends message"
-        text, options = createMessage(message, data)
-        return bot.sendMessage(chatId, text, options)
-    return action
+    def run(self, bot, update):
+        "Apply update to response body"
 
-def sendMessages(*messages):
-    "Returns messages sending function"
-
-    def action(bot, chatId, data):
-        "Sends messages"
-        for message in messages:
-            text, options = createMessage(message, data)
-            bot.sendMessage(chatId, text, options)
-    return action
-
-def sendKeyboard(kb={}):
-    "Returns function that sends keyboard to user"
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.keyboard(chatId, kb)
-    return action
-
-def sendPhoto(photo=None, caption=''):
-    "Returns function that sends photo to user"
-    if not photo:
-        raise ResponseFormatError('Specify photo by URL or File')
-
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.sendPhoto(chatId, photo, caption)
-
-    return action
-
-def sendSticker(stickerId):
-    "Returns stickerSender"
-
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.sendSticker(chatId, stickerId)
-
-    return action
-
-def sendLocation(*latlon):
-    "Returns location sender"
-
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.sendLocation(chatId, *latlon)
-
-    return action
-
-class Response():
-    "Response class"
-
-    __alowedRules = {
-        'sendMessages': sendMessages,
-        'sendMessage': sendMessage,
-        'sendKeyboard': sendKeyboard,
-        'sendSticker': sendSticker,
-        'sendPhoto': sendPhoto,
-        'sendLocation': sendLocation
-    }
-
-    def __addAction(self, actionName, actionArgs):
-        if isinstance(actionArgs, list):
-            action = self.__alowedRules[actionName](*actionArgs)
-        else:
-            action = self.__alowedRules[actionName](actionArgs)
-        self.actions.append(action)
-
-    def __processResponseDescription(self, response):
-        for key in response:
-            if key in self.__alowedRules:
-                self.__addAction(key, response[key])
-            else:
-                raise RuleNameError('Response property: "{}" is invalid'.format(key))
-
-    def __init__(self, responseBody):
-        self.body = responseBody
-        self.actions = []
-
-        responses = flatten(processResponseBody(responseBody))
-
-        for response in responses:
-            self.__processResponseDescription(response)
-
-    def run(self, update, bot):
-        """
-            Runs response actions
-        """
         message = update.get('message', {})
         text = message.get('text')
         sender = message.get('from', {})
@@ -145,5 +52,202 @@ class Response():
             'second_name': sender.get('second_name')
         }
 
+        methodArgs = self.getData(data)
+
+        if not instanceof(methodArgs, [dict, tuple]):
+            methodArgs = [methodArgs]
+
+        sendArguments = [senderId] + list(methodArgs)
+
+        options = self.getOptions()
+        options = {'options': options}
+        getattr(bot, self.method)(*sendArguments, **options)
+
+class Text(Message):
+    "Text response container"
+    # pylint: disable=missing-docstring
+
+    method = 'sendMessage'
+
+    def __init__(self, text, **options):
+        super().__init__(**options)
+        self.text = text
+
+    def getData(self, data):
+        text = self.text
+
+        if self.options.get('interpolate'):
+            text = self.text.format(
+                text=data.get('text'),
+                first_name=data.get('first_name'),
+                username=data.get('username'),
+                second_name=data.get('second_name'))
+
+        return text
+
+    def getOptions(self):
+        options = super().getOptions()
+
+        additional = {
+            'disable_web_page_preview': self.options.get('disableWebPreview', False),
+            'reply_to_message_id': self.options.get('replyId', None)
+        }
+
+        options.update(additional)
+
+        if self.options.get('format'):
+            options.update({'parse_mode': self.options.get('format')})
+
+        if self.options.get('markup'):
+            options.update({'reply_markup': self.options.get('markup')})
+
+        return options
+
+class Keyboard(Text):
+    "Keyboard response container"
+    method = 'sendKeyboard'
+    # pylint: disable=missing-docstring
+
+    def __init__(self, text, keyboard=None, **options):
+        super().__init__(text, **options)
+        self.text = text if keyboard else text.get('text')
+        self.keyboard = keyboard if keyboard else text.get('keyboard', [])
+
+        if isinstance(self.keyboard, list):
+            self.keyboard = {
+                'keyboard': self.keyboard
+            }
+
+        if options.get('autohide') != None:
+            self.keyboard['one_time_keyboard'] = options.get('autohide')
+
+        if options.get('resize') != None:
+            self.keyboard['resize_keyboard'] = options.get('resize')
+
+    def getData(self, data):
+        return super().getData(data), self.keyboard
+
+class Location(Message):
+    "Location response container"
+    # pylint: disable=missing-docstring
+    method = 'sendLocation'
+
+    def __init__(self, lat, lon, **options):
+        super().__init__(**options)
+        self.lat = lat
+        self.lon = lon
+
+    def getData(self, data):
+        return self.lat, self.lon
+
+class Sticker(Message):
+    "Sticker response container"
+    # pylint: disable=missing-docstring
+    def __init__(self, sticker, **options):
+        super().__init__(**options)
+        self.sticker = sticker
+
+    def getData(self, data):
+        return self.sticker
+
+
+class File(Message):
+    "File response container"
+    # pylint: disable=missing-docstring
+
+    method = 'sendDocument'
+
+    def __init__(self, file, caption=None, **options):
+        super().__init__(**options)
+        self.file = file
+        self.caption = caption
+
+    def getData(self, data):
+        return self.file, self.caption
+
+
+class Photo(File):
+    "Photo response container"
+    method = 'sendPhoto'
+
+class Contact(Message):
+    "Contact response container"
+    # pylint: disable=missing-docstring
+    method = 'sendContact'
+
+    def __init__(self, phone, first, last='', **options):
+        super().__init__(**options)
+        self.phone = phone
+        self.first = first
+        self.last = last
+
+    def getData(self, data):
+        return self.phone, self.first, self.last
+
+def getAction(actionType, data):
+    "Calculates action by actionType in response description"
+    # pylint: disable=missing-docstring,unnecessary-lambda,redefined-variable-type
+    response = None
+
+    if actionType == 'text':
+        if isinstance(data, dict):
+            text = data.get('text', '')
+            del data['text']
+            response = Text(text, **data)
+        else:
+            text = data
+            response = Text(text)
+    elif actionType == 'location':
+        if isinstance(data, dict):
+            lat = data.get('lat')
+            lon = data.get('lon')
+            del data['lat']
+            del data['lon']
+            response = Location(lat, lon, **data)
+        elif isinstance(data, list):
+            response = Location(data[0], data[1])
+        else:
+            raise ResponseFormatError('Invalid syntax at Location response')
+
+    def action(response):
+        return lambda bot, update: response.run(bot, update)
+
+    return action(response)
+
+class Response():
+    "Response class"
+
+    __map = {
+        'text': Text,
+        'keyboard': Keyboard,
+        'location': Location,
+        'photo': Photo
+    }
+
+    def __init__(self, responses):
+        self.body = responses
+        self.actions = []
+
+        if not isinstance(responses, list):
+            responses = [responses]
+
+        for response in responses:
+            self.addAction(response)
+
+    def addAction(self, response):
+        "Adds action to actions list"
+        for responseKey in response:
+
+            respBody = response[responseKey]
+
+            if responseKey in self.__map:
+                self.actions.append(getAction(responseKey, respBody))
+            else:
+                raise ResponseFormatError('Invalid response action: "{}"'.format(responseKey))
+
+
+    def run(self, bot, update):
+        "Runs response actions"
+
         for action in self.actions:
-            action(bot, senderId, data)
+            action(bot, update)