+++ bobot/bobot.py
@@ -41,7 +41,7 @@ def caller(method, **kwargs):
     def callerDecorator(dataFunction):
         "Caller decorator"
 
-        def callerFunction(self, chatId, *args):
+        def callerFunction(self, chatId, *args, **options):
             "API Caller"
             argumentsOrder = kwargs.get('arguments')
             requiredArguments = kwargs.get('required')
@@ -56,8 +56,14 @@ def caller(method, **kwargs):
             else:
                 callArgs = dataFunction(*args)
 
+            if isinstance(callArgs, dict):
+                callArgs = [callArgs]
+
             if kwargs.get('static', False) is not True:
                 callArgs[0]['chat_id'] = chatId
+
+            if options.get('options', False):
+                callArgs[0].update(options.get('options'))
 
             return sendFunciton(self.getToken(), *callArgs)
 
@@ -98,24 +104,20 @@ class Bot(object):
         return info.get('result')
 
     @caller('sendMessage',
-            arguments=['text', 'options'],
+            arguments=['text'],
             required={
                 'text': MessageTextEmptyError('Specify message\'s text')
             })
-    def sendMessage(text, options={}):
+    def sendMessage(text):
         "Sends text to user"
 
         data = {
             'text': text
         }
 
-        data.update(options)
+        return data
 
-        return [data]
-
-    @caller('sendSticker', signature={
-        'stricker': 'stickerId'
-    })
+    @caller('sendSticker')
     def sendSticker(stickerId):
         "Sends stricker to user"
 
@@ -123,7 +125,7 @@ class Bot(object):
             'sticker': stickerId
         }
 
-        return [data]
+        return data
 
     @caller('sendPhoto')
     def sendPhoto(photo, caption=None):
@@ -143,13 +145,22 @@ class Bot(object):
     @caller('sendLocation')
     def sendLocation(lat, lon):
         "sen"
-        return [{
+        return {
             'latitude': lat,
             'longitude': lon
-        }]
+        }
+
+    @caller('sendContact')
+    def sendContact(phone, name, secondName=None):
+        "Send contact to user"
+        return {
+            'phone_number': phone,
+            'first_name': name,
+            'secon_name': secondName
+        }
 
     @caller('sendMessage')
-    def keyboard(text, keyboard=None):
+    def sendKeyboard(text, keyboard=None):
         "Sends keyboard to user"
 
         if not keyboard:
@@ -161,10 +172,10 @@ class Bot(object):
 
         board = dumps(keyboard)
 
-        return [{
+        return {
             'text': text,
             'reply_markup': board
-        }]
+        }
 
     def process(self, update):
         """
@@ -210,10 +221,10 @@ class Bot(object):
     @caller('getUpdates', static=True)
     def getUpdates(limit=None, offset=None):
         "Call getUpdates method"
-        return [{
+        return {
             'limit': limit,
             'offset': offset
-        }]
+        }
 
     def register(self, user, registerInfo={'id': 'id'}):
         "Registers client to bot memory"
@@ -234,7 +245,7 @@ class Bot(object):
         if certificate:
             data['certificate'] = certificate
 
-        return [data]
+        return data
 
     def getToken(self):
         "Returns token"